Make timestamps integer

+++ server/src/main/java/com/forum/entities/Category.java
@@ -5,7 +5,7 @@ import com.forum.utils.Time;
 public class Category {
   public String name;
   public String description;
-  public long created_at;
+  public int created_at;
 
   public Category(String name, String description) {
     this.name = name;
